Completed most of NMNN machine learning algorithm.

+++ WraithEngine/src/wraith/library/AI/V7/Score.java
@@ -2,5 +2,5 @@ package wraith.library.AI.V7;
 
 public class Score{
 	public int id;
-	public double value;
+	public long value;
 }
